fix cleanup-unused-users management command

+++ mygpo/data/management/commands/cleanup-unused-users.py
@@ -1,4 +1,5 @@
 from django.core.management.base import BaseCommand
+from django.contrib.auth.models import User
 from mygpo.api.models import Device
 from registration.models import RegistrationProfile
 
@@ -11,27 +12,25 @@ class Command(BaseCommand):
             if not profile.activation_key_expired():
                 continue
 
-            user = profile.user
+            try:
+                user = profile.user
+            except User.DoesNotExist:
+                profile.delete()
+
             try:
                 user_profile = user.get_profile()
                 deleted = user_profile.deleted
             except:
                 deleted = False
 
-            if not user.is_active:
-                if deleted:
-                    continue
+            if not user.is_active and deleted:
 
-                devices = Device.objects.filter(user=user).count()
-                print '%s (%s)' % (user, devices)
+                devices = Device.objects.filter(user=user)
+                print '%s (%s)' % (user, devices.count())
 
-                if devices > 0:
-                    user.is_active = True
-                    profile.activation_key = RegistrationProfile.ACTIVATED
-                    user.save()
-                    profile.save()
+                for device in devices:
+                    device.delete()
 
-                else:
-                    profile.delete()
-                    user.delete()
-
+                profile.delete()
+                user_profile.delete()
+                user.delete()
